+++ docs/changelog.md
@@ -7,6 +7,10 @@
 - 更新Neo4j的Python包依赖，原依赖马上就要废弃了
 - 更改Neo4j的配置，接入本地数据库
 - 本地数据库配置详见：[Neo4j本地部署](https://morethan987.github.io/blog/neo4j-basics/#%E5%AE%89%E8%A3%85)
+- 修改pyproject.toml中的项目名称
+- 在.env.example中添加注释说明
+- 针对图数据库专用语言模型配置产生API_KEY冲突问题进行说明；见test_local.py中的注释说明
+- 解决graph_memory.py中图数据库存储专用模型与外部主配置模型冲突问题
 
 ## 2025-02-14